fixes and filter for GET

+++ src/main/java/com/itmo/wtsc/services/RequestService.java
@@ -1,5 +1,6 @@
 package com.itmo.wtsc.services;
 
+import com.itmo.wtsc.dto.RequestFilter;
 import com.itmo.wtsc.utils.ErrorMessages;
 import com.itmo.wtsc.dto.RequestDto;
 import com.itmo.wtsc.entities.Point;
@@ -14,6 +15,7 @@ import com.itmo.wtsc.utils.enums.UserRole;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.*;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
@@ -52,6 +54,7 @@ public class RequestService {
         request.setPoint(point);
         request.setUser(user);
         request.setStatus(RequestStatus.WAITING);
+        request.setCreatedWhen(LocalDateTime.now());
         pointRepository.save(point);
         requestRepository.save(request);
         return getRequestDto(request, request.getPoint());
@@ -65,6 +68,9 @@ public class RequestService {
 
         switch (user.getRole()) {
             case TOURIST:
+                if (!Objects.equals(user.getId(), request.getUser().getId())) {
+                    throw new DataNotFoundException(String.format(ErrorMessages.REQUEST_NOT_FOUND_ERROR, id));
+                }
                 validateTouristCanChangeRequest(request.getStatus());
                 request.setSize(requestDto.getSize());
                 request.setDescription(requestDto.getDescription());
@@ -79,25 +85,30 @@ public class RequestService {
         return getRequestDto(request, request.getPoint());
     }
 
-    public List<RequestDto> getRequests(List<RequestStatus> statuses) {
-        if (statuses == null) {
-            statuses = Arrays.asList(RequestStatus.WAITING, RequestStatus.IN_PROGRESS, RequestStatus.COMPLETED, RequestStatus.CANCELLED);
-        }
+    public List<RequestDto> getRequests(RequestFilter filter) {
         User user = userService.getAuthenticatedUser();
         Predicate<Request> userPredicate = request -> true;
         if (UserRole.TOURIST.equals(user.getRole())) {
             userPredicate = request -> user.getId().equals(request.getUser().getId());
-            statuses = Arrays.asList(RequestStatus.WAITING, RequestStatus.IN_PROGRESS);
+            filter.setStatuses(Arrays.asList(RequestStatus.WAITING, RequestStatus.IN_PROGRESS));
         }
-        return requestRepository.findRequestByStatusIn(statuses).stream()
+        return requestRepository.findRequestsByStatusInAndDumpTypeInAndSizeLessThanEqualAndCreatedWhenBetween(
+                filter.getStatuses(), filter.getTypes(), filter.getMaxSize(), filter.getStartTime(), filter.getEndTime())
+                .stream()
                 .filter(userPredicate)
                 .map(request -> getRequestDto(request, request.getPoint()))
                 .collect(Collectors.toList());
     }
 
     public void deleteRequest(Integer id) {
+        User user = userService.getAuthenticatedUser();
         Request request = requestRepository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException(String.format(ErrorMessages.REQUEST_NOT_FOUND_ERROR, id)));
+        if (user.getRequests().stream().noneMatch(req -> req.getId().equals(id))) {
+            throw new DataNotFoundException(String.format(ErrorMessages.REQUEST_NOT_FOUND_ERROR, id));
+        }
+        validateTouristCanChangeRequest(request.getStatus());
+        pointRepository.deleteById(request.getPoint().getId());
         requestRepository.deleteById(id);
     }
